Refactor for SDK 54
- Use kebab case
- Remove fake native tab bar on iOS. This makes it easier to migrate to native tab later.
- Use reanimated keyframe animations for wave.
- Make 404 a modal.
