+++ templates/expo-template-default/components/external-link.tsx
