+++ templates/expo-template-default/components/haptic-tab.tsx
