--- templates/expo-template-default/components/ui/TabBarBackground.ios.tsx
@@ -1,26 +1,0 @@
-import { BlurView } from 'expo-blur';
-import { StyleSheet } from 'react-native';
-import { useBottomTabBarHeight } from '@react-navigation/bottom-tabs';
-import { useSafeAreaInsets } from 'react-native-safe-area-context';
-
-export function TabBarBackground() {
-  return (
-    <BlurView
-      // System chrome material automatically adapts to the system's theme
-      // and matches the native tab bar appearance on iOS.
-      tint="systemChromeMaterial"
-      intensity={70}
-      style={StyleSheet.absoluteFill}
-    />
-  );
-}
-
-export function useBottomTabOverflow() {
-  const insets = useSafeAreaInsets();
-  const tabBarHeight = useBottomTabBarHeight();
-
-  return {
-    scrollInsetBottom: tabBarHeight - insets.bottom,
-    paddingBottom: tabBarHeight,
-  };
-}
