--- templates/expo-template-default/components/ui/TabBarBackground.tsx
@@ -1,9 +1,0 @@
-// This is a shim for web and Android where the tab bar is generally opaque.
-export const TabBarBackground = undefined;
-
-export function useBottomTabOverflow() {
-  return {
-    scrollInsetBottom: 0,
-    paddingBottom: 0,
-  }
-}
