+++ templates/expo-template-default/components/ui/icon-symbol.ios.tsx
