+++ templates/expo-template-default/constants/theme.ts
