+++ templates/expo-template-default/hooks/use-color-scheme.ts
