+++ templates/expo-template-default/hooks/use-color-scheme.web.ts
